hub-poller/requirements.txt
httpx==0.27.2
PyYAML==6.0.2
tenacity==9.0.0

